Added .clang-format and applied the clang-format tool to all .c and .h files.
clang-format (http://clang.llvm.org/docs/ClangFormat.html) is a tool for
automatically formatting C/C++/Obj-C code.  This allows a project to define a
formatting style that is applied across the entire project, which can make it
much easier to read.  I've added the .clang-format file at the root of the
project which can be modified to control the styles used.  I've also
formatted all .c and .h files using clang-format and the style in this
commit.