Removed stringification from write_tables.c.
There are two macros, add_struct_member() and add_array_member() that used the
'#' preprocessor operator to stringify the parameter that was passed to them.
Unfortunately, this did not work well with clang-format.  I've stripped out
the stringification and replaced it with the format strings that were being
created by the preprocessor.  The code passes all the tests, but I haven't yet
tried clang-format on the resulting code.  That will be the next step.